pyproject.toml
[project]
name = "kefu-agent"
version = "0.1.0"
description = "Minimal LangChain/LangGraph multimodal customer-service agent for competition submission."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.0",
    "numpy>=2.0.0",
    "pydantic-settings>=2.6.0",
    "sentence-transformers>=2.7.0",
    "tqdm>=4.67.0",
    "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kefu_agent"]

[tool.pytest.ini_options]
pythonpath = ["src", "."]
testpaths = ["tests"]
